fix(dashboard): skip conditions ui schema check
