+++ apps/dashboard/src/components/workflow-editor/steps/configure-step-form.tsx
@@ -1,5 +1,5 @@
+import { zodResolver } from '@hookform/resolvers/zod';
 import {
-  FeatureFlagsKeysEnum,
   IEnvironment,
   StepResponseDto,
   StepTypeEnum,
@@ -10,22 +10,11 @@ import {
 import { AnimatePresence, motion } from 'motion/react';
 import { HTMLAttributes, ReactNode, useCallback, useEffect, useMemo, useState } from 'react';
 import { useForm } from 'react-hook-form';
-import {
-  RiArrowLeftSLine,
-  RiArrowRightSLine,
-  RiCloseFill,
-  RiDeleteBin2Line,
-  RiGuideFill,
-  RiPencilRuler2Fill,
-} from 'react-icons/ri';
+import { RiArrowLeftSLine, RiArrowRightSLine, RiCloseFill, RiDeleteBin2Line, RiPencilRuler2Fill } from 'react-icons/ri';
 import { Link, useNavigate } from 'react-router-dom';
-import { parseJsonLogic } from 'react-querybuilder/parseJsonLogic';
-import { RQBJsonLogic } from 'react-querybuilder';
-import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 
 import { ConfirmationModal } from '@/components/confirmation-modal';
-import { stepSchema } from '@/components/workflow-editor/schema';
 import { PageMeta } from '@/components/page-meta';
 import { Button } from '@/components/primitives/button';
 import { CopyButton } from '@/components/primitives/copy-button';
@@ -34,6 +23,7 @@ import { Input } from '@/components/primitives/input';
 import { Separator } from '@/components/primitives/separator';
 import { SidebarContent, SidebarFooter, SidebarHeader } from '@/components/side-navigation/sidebar';
 import TruncatedText from '@/components/truncated-text';
+import { stepSchema } from '@/components/workflow-editor/schema';
 import { getStepDefaultValues } from '@/components/workflow-editor/step-default-values';
 import {
   flattenIssues,
@@ -50,9 +40,9 @@ import { ConfigureInAppStepPreview } from '@/components/workflow-editor/steps/in
 import { ConfigurePushStepPreview } from '@/components/workflow-editor/steps/push/configure-push-step-preview';
 import { SaveFormContext } from '@/components/workflow-editor/steps/save-form-context';
 import { SdkBanner } from '@/components/workflow-editor/steps/sdk-banner';
+import { SkipConditionsButton } from '@/components/workflow-editor/steps/skip-conditions-button';
 import { ConfigureSmsStepPreview } from '@/components/workflow-editor/steps/sms/configure-sms-step-preview';
 import { UpdateWorkflowFn } from '@/components/workflow-editor/workflow-provider';
-import { useFeatureFlag } from '@/hooks/use-feature-flag';
 import { useFormAutosave } from '@/hooks/use-form-autosave';
 import { INLINE_CONFIGURABLE_STEP_TYPES, STEP_TYPE_LABELS, TEMPLATE_CONFIGURABLE_STEP_TYPES } from '@/utils/constants';
 import { buildRoute, ROUTES } from '@/utils/routes';
@@ -93,7 +83,6 @@ export const ConfigureStepForm = (props: ConfigureStepFormProps) => {
   const { step, workflow, update, environment } = props;
   const navigate = useNavigate();
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-  const isStepConditionsEnabled = useFeatureFlag(FeatureFlagsKeysEnum.IS_STEP_CONDITIONS_ENABLED);
   const supportedStepTypes = [
     StepTypeEnum.IN_APP,
     StepTypeEnum.SMS,
@@ -205,14 +194,6 @@ export const ConfigureStepForm = (props: ConfigureStepFormProps) => {
   const InlineControlValues = STEP_TYPE_TO_INLINE_CONTROL_VALUES[step.type];
 
   const value = useMemo(() => ({ saveForm }), [saveForm]);
-
-  const conditionsCount = useMemo(() => {
-    if (!step.controls.values.skip) return 0;
-
-    const query = parseJsonLogic(step.controls.values.skip as RQBJsonLogic);
-
-    return query.rules.length;
-  }, [step]);
 
   return (
     <>
@@ -335,27 +316,7 @@ export const ConfigureStepForm = (props: ConfigureStepFormProps) => {
             </>
           )}
 
-          {isStepConditionsEnabled && (
-            <>
-              <SidebarContent>
-                <Link to={'./conditions'} relative="path" state={{ stepType: step.type }}>
-                  <Button
-                    variant="secondary"
-                    mode="outline"
-                    className="flex w-full justify-start gap-1.5 text-xs font-medium"
-                  >
-                    <RiGuideFill className="h-4 w-4 text-neutral-600" />
-                    Skip Conditions
-                    <span className="ml-auto flex items-center gap-0.5">
-                      <span>{conditionsCount}</span>
-                      <RiArrowRightSLine className="ml-auto h-4 w-4 text-neutral-600" />
-                    </span>
-                  </Button>
-                </Link>
-              </SidebarContent>
-              <Separator />
-            </>
-          )}
+          <SkipConditionsButton step={step} inSidebar />
 
           {!isSupportedStep && (
             <SidebarContent>
